Refactor: [BADA-289] 코인 관련 코드 리팩토링 #189
